chore(public): setup tests
