+++ example/activists/index.ts
@@ -1,0 +1,1 @@
+export { default as pressure } from './pressure';
