+++ example/jest.config.js
@@ -1,0 +1,4 @@
+module.exports = {
+  // setupFiles: ['<rootDir>/jest.setup.js'],
+  testPathIgnorePatterns: ['<rootDir>/.next/', '<rootDir>/node_modules/']
+}
