+++ example/package.json
@@ -5,7 +5,9 @@
   "scripts": {
     "dev": "next dev",
     "build": "next build",
-    "start": "next start"
+    "start": "next start",
+    "test": "jest --passWithNoTests",
+    "coverage": "jest --coverage --passWithNoTests"
   },
   "dependencies": {
     "apollo-boost": "^0.4.7",
@@ -37,6 +39,7 @@
     "subscriptions-transport-ws": "^0.9.16"
   },
   "devDependencies": {
+    "@babel/core": "^7.10.4",
     "@types/node": "^12.0.12",
     "@types/react": "^16.8.17",
     "@types/react-dom": "16.8.4",
@@ -45,9 +48,11 @@
     "@types/redux-promise": "^0.5.28",
     "@zeit/next-css": "^1.0.1",
     "@zeit/next-sass": "^1.0.1",
+    "babel-jest": "^26.1.0",
     "babel-plugin-inline-import": "^3.0.0",
     "babel-plugin-react-intl": "^7.5.2",
     "babel-plugin-styled-components": "^1.10.7",
+    "jest": "^26.1.0",
     "ts-node": "^8.1.0",
     "typescript": "^3.4.5"
   }
